// EscalafonFrontend/src/app/interfaces/retencion.ts
export interface Retencion {
    /**
     * @type { number } - iRetenId
     */
    iRetenId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { string } - cRetenNumeroDocumento
     */
    cRetenNumeroDocumento?: string;

    /**
     * @type { Date } - dtRetenFechaDocumento
     */
    dtRetenFechaDocumento?: Date;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iTipRetenId
     */
    iTipRetenId?: number;

    /**
     * @type { number } - iComTipMonId
     */
    iComTipMonId?: number;

    /**
     * @type { number } - nRetenMontoTotal
     */
    nRetenMontoTotal?: number;

    /**
     * @type { number } - nRetenNumeroCuotas
     */
    nRetenNumeroCuotas?: number;

    /**
     * @type { number } - nRetenMontoFijoMensual
     */
    nRetenMontoFijoMensual?: number;

    /**
     * @type { number } - nRetenPorcentajeFijoMensual
     */
    nRetenPorcentajeFijoMensual?: number;

    /**
     * @type { number } - iTipBenRetenId
     */
    iTipBenRetenId?: number;

    /**
     * @type { string } - cRetenEntidadNombre
     */
    cRetenEntidadNombre?: string;

    /**
     * @type { string } - cRetenNombreBeneficiario
     */
    cRetenNombreBeneficiario?: string;

    /**
     * @type { string } - cRetenAnotaciones
     */
    cRetenAnotaciones?: string;

}