// EscalafonFrontend/src/app/interfaces/v-retencion.ts
export interface VRetencion {
    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { string } - cRetenNumeroDocumento
     */
    cRetenNumeroDocumento?: string;

    /**
     * @type { Date } - dtRetenFechaDocumento
     */
    dtRetenFechaDocumento?: Date;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iTipRetenId
     */
    iTipRetenId?: number;

    /**
     * @type { number } - iComTipMonId
     */
    iComTipMonId?: number;

    /**
     * @type { number } - nRetenMontoTotal
     */
    nRetenMontoTotal?: number;

    /**
     * @type { number } - nRetenNumeroCuotas
     */
    nRetenNumeroCuotas?: number;

    /**
     * @type { number } - nRetenMontoFijoMensual
     */
    nRetenMontoFijoMensual?: number;

    /**
     * @type { number } - nRetenPorcentajeFijoMensual
     */
    nRetenPorcentajeFijoMensual?: number;

    /**
     * @type { string } - cRetenEntidadNombre
     */
    cRetenEntidadNombre?: string;

    /**
     * @type { number } - iTipBenRetenId
     */
    iTipBenRetenId?: number;

    /**
     * @type { string } - cRetenNombreBeneficiario
     */
    cRetenNombreBeneficiario?: string;

    /**
     * @type { string } - cRetenAnotaciones
     */
    cRetenAnotaciones?: string;

    /**
     * @type { string } - cTipRetenNombre
     */
    cTipRetenNombre?: string;

    /**
     * @type { number } - iRetenId
     */
    iRetenId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

}